+++ azure/src/main/scala/org/broadinstitute/dsde/workbench/azure/models.scala
@@ -44,3 +44,4 @@ final case class AKSToken(value: String) extends AnyVal
 final case class AKSCertificate(value: String) extends AnyVal
 final case class AKSCredentials(server: AKSServer, token: AKSToken, certificate: AKSCertificate)
 final case class ApplicationInsightsName(value: String) extends AnyVal
+final case class BatchAccountName(value: String) extends AnyVal